+++ odoo/local-src/dealsheet_purchase_sourcing/wizards/__init__.py
@@ -1,0 +1,1 @@
+from . import wiz_sale_dealsheet_source