+++ odoo/local-src/sale_purchase_sourcing/__manifest__.py
@@ -2,7 +2,8 @@
 # Author: Simone Orsi <simone.orsi@camptocamp.com>
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
+# TODO Move me to https://github.com/camptocamp/odoo-sale-addons after
+# update 10.3.7
 {
     'name': 'Sale Purchase Procuremnt',
     'summary': """TODO.""",