+++ HISTORY.rst
@@ -24,6 +24,8 @@ Latest (Unreleased)
 * BSIBSO-1032: Change the way analytic accounts / projects are numbered
   the name of the analytic account must be generated as follows AARRR/BBBBCC/DDDDD
 * BSIBSO-1031: move the Procure button from Sale Order to Dealsheet screen
+* BSIBSO-1031 Move procure from sale order to dealsheet
+* BSIBSO-1031 Hide Set cost button on sale order line
 
 **Bugfixes**